Revert layout of hole punch code to use trait
